+++ pyemma/coordinates/data/featurization/featurizer.py
@@ -57,6 +57,8 @@ class MDFeaturizer(Loggable):
             self.topologyfile = topfile
         elif isinstance(topfile, mdtraj.Topology):
             self.topology = topfile
+        elif isinstance(topfile, mdtraj.Trajectory):
+            self.topology = topfile.topology
         else:
             raise ValueError("no valid topfile arg: type was %s, "
                              "but only string or mdtraj.Topology allowed." % type(topfile))
@@ -137,9 +139,15 @@ class MDFeaturizer(Loggable):
         """
         return self.topology.select("backbone and (name C or name CA or name N)")
 
-    def select_Heavy(self):
+    def select_Heavy(self, exclude_symmetry_related=False):
         """
-        Returns the indexes of all heavy atoms (Mass >= 2)
+        Returns the indexes of all heavy atoms (Mass >= 2),
+        optionally excluding symmetry-related heavy atoms.
+        
+        Parameters
+        ----------
+        exclude_symmetry_related : boolean, default=False
+            if True, exclude symmetry-related heavy atoms.
 
         Returns
         -------
@@ -147,8 +155,26 @@ class MDFeaturizer(Loggable):
             array with selected atom indexes
 
         """
-        return self.topology.select("mass >= 2")
-
+        if exclude_symmetry_related:
+            exclusions = []
+        
+            exclusions.append("mass < 2")
+            exclusions.append("(resname == VAL and name == CG)")
+            exclusions.append("(resname == LEU and name == CD)")
+            exclusions.append("(resname == PHE and name == CD) or (resname == PHE and name == CE)")
+            exclusions.append("(resname == TYR and name == CD) or (resname == TYR and name == CE)")
+            exclusions.append("(resname == GLU and name == OD1) or (resname == GLU and name == OD2)")
+            exclusions.append("(resname == ASP and name == OG1) or (resname == ASP and name == OG2)")
+            exclusions.append("(resname == HIS and name == ND1) or (resname == HIS and name == NE2)")
+            exclusions.append("(resname == ARG and name == NH1) or (resname == ARG and name == NH2)")
+            
+            exclusion_string = ' or '.join(exclusions)
+            selection_string = 'not (' + exclusion_string + ')'
+        
+            return self.topology.select(selection_string)
+        else:
+            return self.topology.select("mass >= 2")
+    
     @staticmethod
     def pairs(sel, excluded_neighbors=0):
         """