+++ pyemma/msm/estimators/maximum_likelihood_hmsm.py
@@ -588,7 +588,7 @@ class MaximumLikelihoodHMSM(_Estimator, _HMSM):
     # Model Validation
     ################################################################################
 
-    def cktest(self, mlags=10, conf=0.95, err_est=False, show_progress=True):
+    def cktest(self, mlags=10, conf=0.95, err_est=False, n_jobs=1, show_progress=True):
         """ Conducts a Chapman-Kolmogorow test.
 
         Parameters
@@ -604,6 +604,8 @@ class MaximumLikelihoodHMSM(_Estimator, _HMSM):
             compute errors also for all estimations (computationally expensive)
             If False, only the prediction will get error bars, which is often
             sufficient to validate a model.
+        n_jobs : int, default=1
+            how many jobs to use during calculation
         show_progress : bool, default=True
             Show progressbars for calculation?
 
@@ -629,6 +631,6 @@ class MaximumLikelihoodHMSM(_Estimator, _HMSM):
         from pyemma.msm.estimators import ChapmanKolmogorovValidator
         ck = ChapmanKolmogorovValidator(self, self, _np.eye(self.nstates),
                                         mlags=mlags, conf=conf, err_est=err_est,
-                                        show_progress=show_progress)
+                                        n_jobs=n_jobs, show_progress=show_progress)
         ck.estimate(self._dtrajs_full)
         return ck